+++ micro-migration/src/main/java/software/xdev/micromigration/eclipsestore/TunnelingEmbeddedStorageManager.java
@@ -17,6 +17,7 @@ package software.xdev.micromigration.eclipsestore;
 
 import java.nio.ByteBuffer;
 import java.util.Objects;
+import java.util.function.Consumer;
 import java.util.function.Predicate;
 
 import org.eclipse.serializer.afs.types.AFile;
@@ -26,7 +27,6 @@ import org.eclipse.serializer.persistence.types.PersistenceManager;
 import org.eclipse.serializer.persistence.types.PersistenceRootsView;
 import org.eclipse.serializer.persistence.types.PersistenceTypeDictionaryExporter;
 import org.eclipse.store.storage.embedded.types.EmbeddedStorageManager;
-import org.eclipse.store.storage.exceptions.StorageException;
 import org.eclipse.store.storage.types.Database;
 import org.eclipse.store.storage.types.StorageConfiguration;
 import org.eclipse.store.storage.types.StorageConnection;
@@ -159,7 +159,7 @@ public class TunnelingEmbeddedStorageManager
 	 * {@link EmbeddedStorageManager#close()}
 	 */
 	@Override
-	public void close() throws StorageException
+	public void close()
 	{
 		this.nativeManager.close();
 	}
@@ -415,5 +415,35 @@ public class TunnelingEmbeddedStorageManager
 	public void issueTransactionsLogCleanup()
 	{
 		this.nativeManager.issueTransactionsLogCleanup();
+	}
+	
+	@Override
+	public int markUsedFor(final Object instance)
+	{
+		return this.nativeManager.markUsedFor(instance);
+	}
+	
+	@Override
+	public int unmarkUsedFor(final Object instance)
+	{
+		return this.nativeManager.unmarkUsedFor(instance);
+	}
+	
+	@Override
+	public boolean isUsed()
+	{
+		return this.nativeManager.isUsed();
+	}
+	
+	@Override
+	public int markUnused()
+	{
+		return this.nativeManager.markUnused();
+	}
+	
+	@Override
+	public void accessUsageMarks(final Consumer<? super XGettingEnum<Object>> logic)
+	{
+		this.nativeManager.accessUsageMarks(logic);
 	}
 }