+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 3.0.0
+
+* Added support for EclipseStore v2
+* Having an empty version for the first migration-script is now allowed
+
 ## 2.0.0
 _Major refactoring_
 * Consolidated all previous modules into a single one